docs: update drawer and modal with toast examples
